Machine Learning/requirements.txt
absl-py==2.1.0
asttokens==3.0.0
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.10
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.1
Flask==3.1.0
flatbuffers==24.3.25
fonttools==4.55.3
fsspec==2024.10.0
gast==0.6.0
google-api-core==2.24.0
google-auth==2.36.0
google-cloud-discoveryengine==0.13.4
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
h5py==3.12.1
huggingface-hub==0.26.5
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.7.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.0.2
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
stack-data==0.6.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
termcolor==2.5.0
tf_keras==2.18.0
threadpoolctl==3.5.0
tokenizers==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.0
